+++ ods_tools/oed/common.py
@@ -4,6 +4,8 @@ common static variable and ods_tools exceptions
 from urllib.parse import urlparse
 from pathlib import Path
 import numpy as np
+import pandas as pd
+from enum import Enum
 
 
 class OdsException(Exception):
@@ -87,9 +89,25 @@ DEFAULT_VALIDATION_CONFIG = [
     {'name': 'occupancy_code', 'on_error': 'raise'},
     {'name': 'construction_code', 'on_error': 'raise'},
     {'name': 'country_and_area_code', 'on_error': 'raise'},
+    {'name': 'conditional_requirement', 'on_error': 'raise'},
 ]
 
 OED_PERIL_COLUMNS = ['AccPeril', 'PolPerilsCovered', 'PolPeril', 'CondPeril', 'LocPerilsCovered', 'LocPeril',
                      'ReinsPeril']
 
-BLANK_VALUES = {np.nan, '', None}
+BLANK_VALUES = {np.nan, '', None, pd.NA, pd.NaT}
+
+
+def fill_empty(df, columns, value):
+    if isinstance(columns, str):
+        columns = [columns]
+    for column in columns:
+        if df[column].dtypes.name == 'category' and value not in {None, np.nan}.union(df[column].cat.categories):
+            df[column] = df[column].cat.add_categories(value)
+        df.loc[df[column].isin(BLANK_VALUES), column] = value
+
+
+class UnknownColumnSaveOption(Enum):
+    IGNORE = 1
+    RENAME = 2
+    DELETE = 3