+++ tests/test_ods_package.py
@@ -1,6 +1,8 @@
 import json
+import logging
 import os
 import pathlib
+import pytest
 import shutil
 import urllib
 
@@ -16,8 +18,9 @@ import tempfile
 sys.path.append(sys.path.pop(0))
 
 from ods_tools.main import convert
-from ods_tools.oed import OedExposure, OedSchema, OdsException, ModelSettingSchema, AnalysisSettingSchema
-from ods_tools.oed.common import OdsException
+from ods_tools.oed import OedExposure, OedSchema, OdsException, ModelSettingSchema, AnalysisSettingSchema, OED_TYPE_TO_NAME, UnknownColumnSaveOption
+
+logger = logging.getLogger(__file__)
 
 base_test_path = pathlib.Path(__file__).parent
 
@@ -51,6 +54,10 @@ def _is_non_empty_file(fp):
 
 
 class OdsPackageTests(TestCase):
+    @pytest.fixture(autouse=True)
+    def logging_fixtures(self, caplog):
+        self._caplog = caplog
+
     def test_load_oed_from_config(self):
         config = {'location': base_url + '/SourceLocOEDPiWind.csv',
                   'account': base_url + '/SourceAccOEDPiWind.parquet',
@@ -107,7 +114,7 @@ class OdsPackageTests(TestCase):
             'PortNumber': [1, 1],
             'PortName': ['1', None],
             'AccNumber': [1, 2],
-            'AccName': [1, np.nan],
+            'AccName': [1, ''],
             'LocNumber': [1, 2],
             'COUNTRYCODE ': ['GB', 'FR'],
             'LocPerilsCovered': 'WTC',
@@ -146,6 +153,40 @@ class OdsPackageTests(TestCase):
             # check csv stream is read
             location = exposure.location.dataframe
             self.assertTrue(isinstance(location, pd.DataFrame))
+
+    def test_parquet_and_csv_return_same_df(self):
+        with tempfile.TemporaryDirectory() as tmp_run_dir:
+            # read_parquet needs stream with seek method which urllib.request.urlopen doesn't have
+            configs_from_file = {}
+            configs_from_stream = {}
+            for extension in ['csv', 'parquet']:
+                configs_from_file[extension] = {'use_field': True}
+                configs_from_stream[extension] = {'use_field': True}
+                for oed_type, oed_name in OED_TYPE_TO_NAME.items():
+                    configs_from_file[extension][oed_name] = os.path.join(tmp_run_dir, f'Source{oed_type}OEDPiWind.{extension}')
+                    with open(configs_from_file[extension][oed_name], 'wb') as acc_parquet:
+                        acc_parquet.write(urllib.request.urlopen(base_url + f'/Source{oed_type}OEDPiWind.{extension}').read())
+                    configs_from_stream[extension][oed_name] = open(os.path.join(tmp_run_dir, f'Source{oed_type}OEDPiWind.{extension}'), 'rb')
+
+            exposure_csv_file = OedExposure(**configs_from_file['csv'])
+            exposure_parquet_file = OedExposure(**configs_from_file['parquet'])
+            exposure_csv_stream = OedExposure(**configs_from_stream['csv'])
+            exposure_parquet_stream = OedExposure(**configs_from_stream['parquet'])
+
+            pd.testing.assert_frame_equal(exposure_csv_file.location.dataframe, exposure_parquet_file.location.dataframe, check_categorical=False)
+            pd.testing.assert_frame_equal(exposure_csv_file.account.dataframe, exposure_parquet_file.account.dataframe, check_categorical=False)
+            pd.testing.assert_frame_equal(exposure_csv_file.ri_info.dataframe, exposure_parquet_file.ri_info.dataframe, check_categorical=False)
+            pd.testing.assert_frame_equal(exposure_csv_file.ri_scope.dataframe, exposure_parquet_file.ri_scope.dataframe, check_categorical=False)
+
+            pd.testing.assert_frame_equal(exposure_csv_file.location.dataframe, exposure_csv_stream.location.dataframe, check_categorical=False)
+            pd.testing.assert_frame_equal(exposure_csv_file.account.dataframe, exposure_csv_stream.account.dataframe, check_categorical=False)
+            pd.testing.assert_frame_equal(exposure_csv_file.ri_info.dataframe, exposure_csv_stream.ri_info.dataframe, check_categorical=False)
+            pd.testing.assert_frame_equal(exposure_csv_file.ri_scope.dataframe, exposure_csv_stream.ri_scope.dataframe, check_categorical=False)
+
+            pd.testing.assert_frame_equal(exposure_csv_file.location.dataframe, exposure_parquet_stream.location.dataframe, check_categorical=False)
+            pd.testing.assert_frame_equal(exposure_csv_file.account.dataframe, exposure_parquet_stream.account.dataframe, check_categorical=False)
+            pd.testing.assert_frame_equal(exposure_csv_file.ri_info.dataframe, exposure_parquet_stream.ri_info.dataframe, check_categorical=False)
+            pd.testing.assert_frame_equal(exposure_csv_file.ri_scope.dataframe, exposure_parquet_stream.ri_scope.dataframe, check_categorical=False)
 
     def test_load_oed_from_stream__detect_type(self):
         with tempfile.TemporaryDirectory() as tmp_run_dir:
@@ -262,7 +303,7 @@ class OdsPackageTests(TestCase):
                         os.path.isfile(pathlib.Path(tmp_run_dir, folder, f'Source{oed_name}OEDPiWind.parquet')))
 
     def test_validation_raise_exception(self):
-        config = {'location': base_url + '/SourceLocOEDPiWind.csv',
+        config = {'location': base_url + '/SourceLocOEDPiWind10.csv',
                   'account': base_url + '/SourceAccOEDPiWind.csv',
                   'ri_info': base_url + '/SourceReinsInfoOEDPiWind.csv',
                   'ri_scope': base_url + '/SourceReinsScopeOEDPiWind.csv',
@@ -270,11 +311,20 @@ class OdsPackageTests(TestCase):
         exposure = OedExposure(**config)
 
         # create invalid data
-        exposure.location.dataframe['BuildingTIV'][0:5] = -1
-        exposure.account.dataframe['LayerParticipation'][1:2] = 2
+        exposure.location.dataframe.loc[0:5, ('BuildingTIV',)] = -1
+        exposure.location.dataframe['LocLimitType6All'] = 0
+        exposure.location.dataframe['LocLimit6All'] = 10000
+        exposure.location.dataframe['CondTag'] = '1'  # check CR only apply to relevant source type
+        exposure.account.dataframe.loc[1:2, ('LayerParticipation',)] = 2
         exposure.ri_info.dataframe['ReinsPeril'] = 'FOOBAR'
 
-        self.assertRaises(OdsException, exposure.check)
+        with pytest.raises(OdsException) as e:
+            exposure.check()
+            self.assertTrue("column 'BuildingTIV' has values outside range." in e.msg)
+            self.assertTrue("column 'LayerParticipation' has values outside range." in e.msg)
+            self.assertTrue("ReinsPeril has invalid perils." in e.msg)
+            self.assertTrue("Conditionally required column missing" in e.msg and 'LocPeril' in e.msg)
+            self.assertfalse('CondPriority' in e.msg)
 
     # load non utf-8 file
     def test_load_non_utf8(self):
@@ -309,7 +359,7 @@ class OdsPackageTests(TestCase):
                 'check_oed': True})
             pd.testing.assert_series_equal(exposure.location.dataframe['LocNumber'],
                                            location_df['LocNumberAlias'].astype(str).astype('category'),
-                                           check_names=False)
+                                           check_names=False, check_categorical=False)
 
     def test_load_exposure_from_different_directory(self):
         """
@@ -354,8 +404,8 @@ class OdsPackageTests(TestCase):
             with tempfile.TemporaryDirectory() as tmp_dir:
                 abs_dir = pathlib.Path(tmp_dir, "abs")
                 abs_dir.mkdir()
-                with urllib.request.urlopen(base_url + '/SourceLocOEDPiWind10Currency.csv') as response,\
-                        open(pathlib.Path(tmp_dir, 'SourceLocOEDPiWind10Currency.csv'), 'wb') as out_file:
+                with urllib.request.urlopen(base_url + '/SourceLocOEDPiWind10.csv') as response,\
+                        open(pathlib.Path(tmp_dir, 'SourceLocOEDPiWind10.csv'), 'wb') as out_file:
                     shutil.copyfileobj(response, out_file)
                 with urllib.request.urlopen(base_url + '/SourceAccOEDPiWind.csv') as response,\
                         open(pathlib.Path(abs_dir, 'SourceAccOEDPiWind.csv'), 'wb') as out_file:
@@ -363,7 +413,7 @@ class OdsPackageTests(TestCase):
 
                 os.chdir(tmp_dir)
                 original_exposure = OedExposure(**{
-                    'location': 'SourceLocOEDPiWind10Currency.csv',  # relative path
+                    'location': 'SourceLocOEDPiWind10.csv',  # relative path
                     'account': str(abs_dir) + '/SourceAccOEDPiWind.csv', })   # absolute path
                 original_exposure.check()
         finally:
@@ -383,6 +433,39 @@ class OdsPackageTests(TestCase):
                 'LocCurrency': ['GBP', 'EUR']}).to_csv(loc_path)
             oed = OedExposure(**{'location': loc_path})
             assert oed.location.dataframe['CountryCode'][0] == 'NA'
+
+    def test_unknown_column_management(self):
+        with tempfile.TemporaryDirectory() as tmp_dir:
+            loc_path = pathlib.Path(tmp_dir, 'location.csv')
+            df = pd.DataFrame({
+                'PortNumber': [1, 1],
+                'AccNumber': [1, 2],
+                'LocNumber': [1, 2],
+                'CountryCode': ['NA', 'FR'],
+                'LocPerilsCovered': 'WTC',
+                'BuildingTIV': ['1000', '20000'],
+                'ContentsTIV': [0, 0],
+                'LocCurrency': ['GBP', 'EUR'],
+                'ToRename': [1, 2],
+                'ToDelete': [3, 4],
+                'ToIgnore': [5, 6],
+                'default_rename': [7, 8],
+            })
+            oed = OedExposure(**{'location': df})
+            save_option = {
+                'ToRename': UnknownColumnSaveOption.RENAME,
+                'ToDelete': UnknownColumnSaveOption.DELETE,
+                'ToIgnore': UnknownColumnSaveOption.IGNORE,
+                'default': UnknownColumnSaveOption.RENAME,
+            }
+
+            oed.save(tmp_dir, unknown_columns=save_option)
+            oed_saved = OedExposure(**{'location': loc_path})
+
+            assert 'ToIgnore' in oed_saved.location.dataframe.columns
+            assert 'FlexiLocToRename' in oed_saved.location.dataframe.columns
+            assert 'ToDelete' not in oed_saved.location.dataframe.columns
+            assert 'FlexiLocdefault_rename' in oed_saved.location.dataframe.columns
 
     def test_setting_schema_analysis__is_valid(self):
         file_name = 'analysis_settings.json'
@@ -489,3 +572,19 @@ class OdsPackageTests(TestCase):
 
             with self.assertRaises(OdsException):
                 ods_model_setting.validate(settings_dict)
+
+    def test_empty_dataframe_logged(self):
+        loc_df = pd.DataFrame({
+            'PortNumber': [],
+            'AccNumber': [],
+            'LocNumber': [],
+            'CountryCode': [],
+            'LocPerilsCovered': [],
+            'BuildingTIV': [],
+            'ContentsTIV': [],
+            'LocCurrency': []})
+        with self._caplog.at_level(logging.INFO):
+            oed = OedExposure(**{'location': loc_df, 'use_field': True})
+            oed.check()
+        assert 'location ' in self._caplog.text
+        assert 'is empty' in self._caplog.text