+++ CHANGELOG.rst
@@ -1,6 +1,21 @@
 ODS_Tools Changelog
 ===================
 
+
+`3.1.0`_
+ ---------
+* [#33](https://github.com/OasisLMF/ODS_Tools/pull/32) - Consistant managment of blank values
+* [#35](https://github.com/OasisLMF/ODS_Tools/pull/36) - Bug in Analysis settings compatibility map
+* [#40](https://github.com/OasisLMF/ODS_Tools/pull/40) - Replace hardcoded packages, with find_packages
+* [#41](https://github.com/OasisLMF/ODS_Tools/pull/41) - fix check category
+* [#42](https://github.com/OasisLMF/ODS_Tools/pull/43) - ods_tools check crashes on empty location file
+* [#44](https://github.com/OasisLMF/ODS_Tools/pull/44) - add check for conditional requirement
+* [#46](https://github.com/OasisLMF/ODS_Tools/pull/46) - add option to manage unknown column when saving files
+* [#27](https://github.com/OasisLMF/ODS_Tools/pull/27) - fix read_csv only read needed line when reading header
+* [#30](https://github.com/OasisLMF/ODS_Tools/pull/30) - Release/3.0.6
+* [#31](https://github.com/OasisLMF/ODS_Tools/pull/31) - Add Platform testing to ODS-tools github actions
+.. _`3.1.0`:  https://github.com/OasisLMF/ODS_Tools/compare/3.0.7...3.1.0
+
 `3.0.6`_
  ---------
 * [#21](https://github.com/OasisLMF/ODS_Tools/pull/22) - LocPeril raise validation error when Blank but it should be allowed according to schema
@@ -12,11 +27,11 @@ ODS_Tools Changelog
 
 `3.0.5`_
  ---------
-* [#5](https://github.com/OasisLMF/ODS_Tools/pull/1) - Separate the ods-tools code from the data standard 
-* [#8](https://github.com/OasisLMF/ODS_Tools/pull/9) - Parquet OED ReinsScope incorrectly fail validation 
-* [#12](https://github.com/OasisLMF/ODS_Tools/pull/13) - Analysis Settings in the  OasisPlatform OpenAPI schema fails build check 
+* [#5](https://github.com/OasisLMF/ODS_Tools/pull/1) - Separate the ods-tools code from the data standard
+* [#8](https://github.com/OasisLMF/ODS_Tools/pull/9) - Parquet OED ReinsScope incorrectly fail validation
+* [#12](https://github.com/OasisLMF/ODS_Tools/pull/13) - Analysis Settings in the  OasisPlatform OpenAPI schema fails build check
 * [#15](https://github.com/OasisLMF/ODS_Tools/pull/15) - Release/3.0.4
-* [#17](https://github.com/OasisLMF/ODS_Tools/pull/18) - Loading blank fields from DataFrame creates string value 'None' 
+* [#17](https://github.com/OasisLMF/ODS_Tools/pull/18) - Loading blank fields from DataFrame creates string value 'None'
 * [#19](https://github.com/OasisLMF/ODS_Tools/pull/19) - support stream as Oed Source
 * [#20](https://github.com/OasisLMF/ODS_Tools/pull/20) - detect stream_obj type if not set
 * [#21](https://github.com/OasisLMF/ODS_Tools/pull/22) - LocPeril raise validation error when Blank but it should be allowed according to schema
@@ -24,8 +39,8 @@ ODS_Tools Changelog
 
 `3.0.4`_
  ---------
-* [#5](https://github.com/OasisLMF/ODS_Tools/pull/1) - Separate the ods-tools code from the data standard 
-* [#12](https://github.com/OasisLMF/ODS_Tools/pull/13) - Analysis Settings in the  OasisPlatform OpenAPI schema fails build check 
-* [#8](https://github.com/OasisLMF/ODS_Tools/pull/9) - Parquet OED ReinsScope incorrectly fail validation 
+* [#5](https://github.com/OasisLMF/ODS_Tools/pull/1) - Separate the ods-tools code from the data standard
+* [#12](https://github.com/OasisLMF/ODS_Tools/pull/13) - Analysis Settings in the  OasisPlatform OpenAPI schema fails build check
+* [#8](https://github.com/OasisLMF/ODS_Tools/pull/9) - Parquet OED ReinsScope incorrectly fail validation
 .. _`3.0.4`:  https://github.com/OasisLMF/ODS_Tools/compare/3.0.3...3.0.4